immediate fixes for AA. ops and turnover costs pounds to pence, large number handling

+++ app/connectors/SubmissionConnector.scala
@@ -50,8 +50,8 @@ trait SubmissionConnector {
       s"$hasPercentageWithMasters/has-ten-year-plan/$hasTenYearPlan")
   }
 
-  def checkLifetimeAllowanceExceeded(hadPrevRFI: Boolean, isKi: Boolean, previousInvestmentSchemesTotal: Int,
-                                     proposedAmount: Int)
+  def checkLifetimeAllowanceExceeded(hadPrevRFI: Boolean, isKi: Boolean, previousInvestmentSchemesTotal: Long,
+                                     proposedAmount: Long)
                                     (implicit hc: HeaderCarrier): Future[Option[Boolean]] = {
 
     http.GET[Option[Boolean]](s"$serviceUrl/investment-tax-relief/lifetime-allowance/lifetime-allowance-checker/had-previous-rfi/" +
